Weirdest Bug in AddRoundKey + KeyScheduling + Encrypt
I've encountered a really strange bug:
The free function AESUtil.encrypt takes the plaintext as an argument,
modifies it while encrypting it and finally exits returning to the
caller function. One would expect that changes to the `plaintext`
parameter would propagate through the return of encrypt. But that's not
happening. Instead we observe that the `plaintext` Parameter gets
changed completely to something else out of thin air. It has become some
other array also declared in the AESTest class.

To make things even stranger: I've hacked my way around this bug by
returning the `plaintext` parameter. This should in theory make no
difference and the returned parameter and the `plaintext` parameter
should be identical. But as we observed the returned parameter is
correct and the other one not. idk pls send help :^)